refactor(capabilities): split the http server runtime module into submodules
